requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git
transformers>=4.40.0
datasets>=2.10.0
pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.2.0
matplotlib>=3.7.0
seaborn>=0.12.2
tqdm>=4.66.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
wandb>=0.16.0
